custom-elements/concepts/custom-elements-type-allowed-chars.html

<!DOCTYPE html>
<html>
<head>
<title>Chars allowed for custom element type </title>
<meta name="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<meta name="assert" content="The custom element type identifies a custom element interface and is a sequence of characters that must match the NCName production">
<link rel="help" href="https://dvcs.w3.org/hg/webcomponents/raw-file/default/spec/custom/index.html#concepts">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src='../testcommon.js'></script>
<link rel="stylesheet" href="/resources/testharness.css">
</head>
<body>
<div id="log"></div>
<script type="text/javascript">

test(function() {
    for (var i = 0; i < EXTENDER_CHARS.length; i++) {
        var doc = newHTMLDocument();
        
        var name = 'x-' + String.fromCharCode(EXTENDER_CHARS[i]);
        var thrown = false;    
        
        try {
            doc.register(name);
        } catch (e) {
            thrown = true;
        }
        
        assert_false(thrown, 'Exception should not be thrown in case of attempt ' +
                'to register a custom element with a name \'' + name + '\' (char code U+' + 
                        EXTENDER_CHARS[i].toString(16) + ')');
    }
}, 'Register custom element types with extender char');


test(function() {
    for (var i = 0; i < COMBINING_CHARS.length; i++) {
        var doc = newHTMLDocument();
        
        var name = 'x-' + String.fromCharCode(COMBINING_CHARS[i]);
        var thrown = false;    
        
        try {
            doc.register(name);
        } catch (e) {
            thrown = true;
        }
        
        assert_false(thrown, 'Exception should not be thrown in case of attempt ' +
                'to register a custom element with a name \'' + name + '\' (char code U+' + 
                        COMBINING_CHARS[i].toString(16) + ')');
    }
}, 'Register custom element types with combining char');


test(function() {
    
    for (var i = 0; i < BASE_CHARS_SINGLE.length; i++) {
        var doc = newHTMLDocument();
        
        var name = 'x-' + String.fromCharCode(BASE_CHARS_SINGLE[i]);
        var thrown = false;    
        
        try {
            doc.register(name);
        } catch (e) {
            thrown = true;
        }
        
        assert_false(thrown, 'Exception should not be thrown in case of attempt ' +
                'to register a custom element with a name \'' + name + '\' (char code U+' + 
                BASE_CHARS_SINGLE[i].toString(16) + ')');
    }
}, 'Register custom element types with single base chars');


test(function() {
    
    var k = 0;
    for (var i = 0; i < BASE_CHARS_RANGES.length - 1; i += 2) {
        for (var j = 0; BASE_CHARS_RANGES[i]+j <= BASE_CHARS_RANGES[i+1]; j++ ){
            var doc = newHTMLDocument();
            
            var charCode = BASE_CHARS_RANGES[i]+j;
            var name = 'x-' + String.fromCharCode(charCode) + '-' + charCode.toString(16);
            
            var thrown = false;    
            
            try {
                doc.register(name);
            } catch (e) {
                thrown = true;
            }
            
            assert_false(thrown, 'Exception should not be thrown in case of attempt ' +
                    'to register a custom element with a name \'' + name + '\' (char code U+' + 
                    charCode.toString(16) + ')');
            k++;
        }
    }
}, 'Register custom element types with base chars ranges');


test(function() {
    
    for (var i = 0; i < IDEOGRAPHIC_CHARS_SINGLE.length; i++) {
        var doc = newHTMLDocument();
        
        var name = 'x-' + String.fromCharCode(IDEOGRAPHIC_CHARS_SINGLE[i]);
        var thrown = false;    
        
        try {
            doc.register(name);
        } catch (e) {
            thrown = true;
        }
        
        assert_false(thrown, 'Exception should not be thrown in case of attempt ' +
                'to register a custom element with a name \'' + name + '\' (char code U+' + 
                IDEOGRAPHIC_CHARS_SINGLE[i].toString(16) + ')');
    }
}, 'Register custom element types with single ideographic chars');


test(function() {
    
    var k = 0;
    for (var i = 0; i < IDEOGRAPHIC_CHARS_RANGES.length - 1; i += 2) {
        for (var j = 0; IDEOGRAPHIC_CHARS_RANGES[i]+j <= IDEOGRAPHIC_CHARS_RANGES[i+1]; j++ ){
            var doc = newHTMLDocument();
            
            var charCode = IDEOGRAPHIC_CHARS_RANGES[i]+j;
            var name = 'x-' + String.fromCharCode(charCode) + '-' + charCode.toString(16);
            
            var thrown = false;    
            
            try {
                doc.register(name);
            } catch (e) {
                thrown = true;
            }
            
            assert_false(thrown, 'Exception should not be thrown in case of attempt ' +
                    'to register a custom element with a name \'' + name + '\' (char code U+' + 
                    charCode.toString(16) + ')');
            k++;
        }
    }
}, 'Register custom element types with ideographic chars ranges');


test(function() {
    
    var k = 0;
    for (var i = 0; i < DIGIT_CHARS_RANGES.length - 1; i += 2) {
        for (var j = 0; DIGIT_CHARS_RANGES[i]+j <= DIGIT_CHARS_RANGES[i+1]; j++ ){
            var doc = newHTMLDocument();
            
            var charCode = DIGIT_CHARS_RANGES[i]+j;
            var name = 'x-' + String.fromCharCode(charCode) + '-' + charCode.toString(16);
            
            var thrown = false;    
            
            try {
                doc.register(name);
            } catch (e) {
                thrown = true;
            }
            
            assert_false(thrown, 'Exception should not be thrown in case of attempt ' +
                    'to register a custom element with a name \'' + name + '\' (char code U+' + 
                    charCode.toString(16) + ')');
            k++;
        }
    }
}, 'Register custom element types with digit chars');

</script>
</body>
</html>
